third_party/WebKit/LayoutTests/crypto/encrypt-decrypt-expected.txt

CONSOLE ERROR: Algorithm: AES-CBC: AesCbcParams: iv: Not an ArrayBufferView
CONSOLE ERROR: Algorithm: AES-CBC: AesCbcParams: iv: Missing required property
CONSOLE ERROR: Algorithm: AES-CBC: AesCbcParams: iv: Not an ArrayBufferView
CONSOLE ERROR: Algorithm: AES-CBC: AesCbcParams: iv: Must be 16 bytes
CONSOLE ERROR: Algorithm: AES-CTR: AesCtrParams: counter: Not an ArrayBufferView
CONSOLE ERROR: Algorithm: AES-CTR: AesCtrParams: counter: Missing required property
CONSOLE ERROR: Algorithm: AES-CTR: AesCtrParams: length: Missing required property
CONSOLE ERROR: key.algorithm does not match that of operation
CONSOLE ERROR: key.algorithm does not match that of operation
CONSOLE ERROR: Algorithm: AES-CTR: AesCtrParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesCtrParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-CTR: AesCtrParams: length: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: iv: Missing required property
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: iv: Not an ArrayBufferView
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: iv: Not an ArrayBufferView
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: additionalData: Not an ArrayBufferView
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: tagLength: Is not a number
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: tagLength: Outside of numeric range
CONSOLE ERROR: Algorithm: AES-GCM: AesGcmParams: tagLength: Outside of numeric range
CONSOLE ERROR: key.algorithm does not match that of operation
CONSOLE ERROR: key.usages does not permit this operation
Tests cypto.subtle.encrypt and crypto.subtle.decrypt

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: Failed to import key of length 0 bytes
PASS: Failed to import key of length 1 bytes
PASS: Failed to import key of length 15 bytes
PASS: Failed to import key of length 17 bytes
PASS: Failed to import key of length 31 bytes
PASS: Failed to import key of length 33 bytes
PASS: Failed to import key of length 23 bytes
PASS: Failed to import key of length 25 bytes
PASS: Failed to import key of length 64 bytes
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-CBC'
PASS tmpKey.algorithm.length is 128
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-CBC'
PASS tmpKey.algorithm.length is 192
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-CBC'
PASS tmpKey.algorithm.length is 256
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-CBC'
PASS tmpKey.algorithm.length is 128
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-GCM'
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-GCM'
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-GCM'
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS tmpKey.type is 'secret'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is 'AES-GCM'
PASS tmpKey.usages.join(',') is 'decrypt,encrypt'
PASS: Encryption should be [7649abac8119b246cee98e9b12e9197d5086cb9b507219ee95db113a917678b273bed6b8e3c1743b7116e69e222295163ff1caa1681fac09120eca307586e1a78cb82807230e1321d3fae00d18cc2012] and was
PASS: Encryption should be [4f021db243bc633d7178183a9fa071e8b4d9ada9ad7dedf4e5e738763f69145a571b242012fb7ae07fa9baac3df102e0288c6f9ec554652e50ab55e121f099ae] and was
PASS: Encryption should be [f58c4c04d6e5f1ba779eabfb5f7bfbd69cfc4e967edb808d679f777bc6702c7d39f23369a9d9bacfa530e26304231461c9aaf02a6a54e9e242ccbf48c59daca6] and was
PASS: Encryption should be [8518b8878d34e7185e300d0fcc426396] and was
PASS: Encryption should be [72ac8493e3a5228b5d130a69d2510e42] and was
PASS: Encryption should be [751f3098d59cf4ea1d2fb0853bde1c] and was
PASS: Encryption should be [bdd2ed6c66fa087dce617d7fd1ff6d93ba82e49c55a22ed02ca67da4ec6f] and was
PASS: Encryption should be [c4489fa64a6edf80e7e6a3b8855bc37c772ee7de0f91a981c36c93a35c88] and was
PASS: Decryption succeeded
PASS: decrypting failed. ciphertext length: 0
PASS: decrypting failed. ciphertext length: 79
PASS: decrypting failed. ciphertext length: 64
PASS: decrypting failed. ciphertext length: 1
PASS: decrypting failed. ciphertext length: 15
PASS: decrypting failed. ciphertext length: 16
PASS: decrypting failed. ciphertext length: 17
PASS: Decryption should be [6bc1bee22e409f96e93d7e117393172aae2d8a571e03ac9c9eb76fac45af8e5130c81c46a35ce411e5fbc1191a0a52eff69f2445df4f9b17ad2b417be66c3710] and was
PASS: Decryption should be [6bc1bee22e409f96e93d7e117393172aae2d8a571e03ac9c9eb76fac45af8e5130c81c46a35ce411e5fbc1191a0a52eff6] and was
PASS: Decryption should be [6bc1bee22e409f96e93d7e117393172aae2d8a571e03ac9c9eb76fac45af8e5130c81c46a35ce411e5fbc1191a0a52eff69f2445df4f9b17ad2b417be6] and was
PASS: Decryption should be [] and was
PASS: Decryption should be [] and was
PASS: Decryption should be [] and was
PASS: Decryption should be [bfaaaea3880d72d4378561e2597a9b35] and was
PASS: Decryption should be [d30b02c343487105219d6fa080acc743] and was
PASS crypto.subtle.encrypt({name: 'bogus'}, null, data) threw exception TypeError: Failed to execute 'encrypt' on 'SubtleCrypto': Invalid key argument.
PASS: 'crypto.subtle.encrypt({name: 'AES-CBC', iv: null}, keys.aesCbc, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CBC'}, keys.aesCbc, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CBC', iv: 3}, keys.aesCbc, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CBC', iv: new Uint8Array(0)}, keys.aesCbc, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: null}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR'}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: new Uint8Array(0)}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: new Uint8Array(16), length: 0}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: new Uint8Array(16), length: 18}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: new Uint8Array(16), length: 256}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: new Uint8Array(16), length: -3}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-CTR', counter: new Uint8Array(16), length: Infinity}, keys.aesCtr, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm'}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm', iv: 3}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm', iv: 'foo'}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: '5'}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: new Uint8Array(1), tagLength: 'foo'}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: new Uint8Array(1), tagLength: -1}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt({name: 'AES-gcm', iv: new Uint8Array(16), additionalData: new Uint8Array(1), tagLength: 8000}, keys.aesGcm, data)' rejected with null
PASS: 'crypto.subtle.encrypt(aesCbc, keys.hmacSha1, data)' rejected with null
PASS: 'crypto.subtle.encrypt(aesCbc, keys.aesCbcJustDecrypt, data)' rejected with null
PASS successfullyParsed is true

TEST COMPLETE
